+++ src/rocdecode-host/avcodec/avcodec_videodecoder.h
@@ -164,4 +164,5 @@ private:
     AVFormatContext * formatContext = nullptr;
     AVInputFormat * inputFormat = nullptr;
 
+    RocDecLogger logger_;
 };