+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Full documentation for rocDecode is available at [https://rocm.docs.amd.com/projects/rocDecode/en/latest/](https://rocm.docs.amd.com/projects/rocDecode/en/latest/)
 
+## (Unreleased) rocDecode 1.5.1
+
+### Changed
+* Updated libdrm path configuration and libva version requirements for ROCm and TheRock platforms
+
+### Added
+* Logging control. Message output from the core components is now controlled by the logging level threshold, which can be set by an environment variable or other methods.
+
 ## rocDecode 1.4.0 for ROCm 7.1.0
 
 ### Added